requirements.txt
presidio_analyzer == 2.2.353
presidio-anonymizer == 2.2.353
langchain == 0.1.9
langchain-experimental == 0.0.52
Faker == 23.2.1
flask == 2.2.3
flask-cors == 4.0.0
Werkzeug==2.2.2 
boto3==1.34.74
pandas
python-dotenv==1.0.1